+++ example/lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:example/data.dart';
+import 'package:example/multi_select_field_examples/multi_select_field_demo.dart';
+import 'package:example/multi_select_field_examples/multi_select_options_control.dart';
 import 'package:flutter/material.dart';
 import 'package:select_field/select_field.dart';
 
@@ -36,7 +38,7 @@ class _MyHomePageState extends State<MyHomePage> {
     return Scaffold(
       appBar: AppBar(
         title: const Text(
-          'Select Field',
+          'Multi Select Field',
           style: TextStyle(
             color: Colors.white,
             fontWeight: FontWeight.bold,
@@ -67,236 +69,11 @@ class _MyHomePageState extends State<MyHomePage> {
             // ),
             // CustomMenuSelectField(options: options),
             // const SizedBox(height: 32),
-            MultiSelectDemo<String>(options: options),
+            // MultiSelectOptionsControl<String>(options: options),
+            MultiSelectFieldDemo<String>(options: options)
           ],
         ),
       ),
-    );
-  }
-}
-
-class StyledSelectField extends StatelessWidget {
-  final List<Option<String>> options;
-
-  const StyledSelectField({
-    super.key,
-    required this.options,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SelectField<String>(
-      options: options,
-      initialOption: Option<String>(
-        label: fruitOptions[0],
-        value: fruitOptions[0],
-      ),
-      menuPosition: MenuPosition.below,
-      onTextChanged: (value) => debugPrint(value),
-      onOptionSelected: (option) => debugPrint(option.toString()),
-      inputStyle: const TextStyle(),
-      menuDecoration: MenuDecoration(
-        margin: const EdgeInsets.only(top: 8),
-        height: 365,
-        width: null,
-        alignment: MenuAlignment.center,
-        backgroundDecoration: BoxDecoration(
-          color: Colors.green[100],
-          borderRadius: BorderRadius.circular(12),
-          boxShadow: [
-            BoxShadow(
-              offset: const Offset(1, 1),
-              color: Colors.brown[300]!,
-              blurRadius: 3,
-            ),
-          ],
-        ),
-        animationDuration: const Duration(milliseconds: 400),
-        buttonStyle: TextButton.styleFrom(
-          fixedSize: const Size(double.infinity, 60),
-          backgroundColor: Colors.green[100],
-          alignment: Alignment.centerLeft,
-          padding: const EdgeInsets.all(16),
-          shape: const RoundedRectangleBorder(),
-          textStyle: const TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w600,
-            color: Colors.green,
-          ),
-        ),
-        separatorBuilder: (context, index) => Container(
-          height: 1,
-          width: double.infinity,
-          color: Colors.green,
-        ),
-      ),
-    );
-  }
-}
-
-class CustomMenuSelectField<String> extends StatefulWidget {
-  final List<Option<String>> options;
-
-  const CustomMenuSelectField({
-    super.key,
-    required this.options,
-  });
-
-  @override
-  State<CustomMenuSelectField<String>> createState() =>
-      _CustomMenuSelectFieldState<String>();
-}
-
-class _CustomMenuSelectFieldState<String>
-    extends State<CustomMenuSelectField<String>> {
-  Option<String>? selectedOption;
-  final SelectFieldMenuController<String> menuController =
-      SelectFieldMenuController();
-
-  @override
-  void initState() {
-    super.initState();
-    menuController.addListener(() {
-      setState(() {});
-    });
-  }
-
-  @override
-  void dispose() {
-    menuController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SelectField<String>(
-      menuController: menuController,
-      options: widget.options,
-      menuPosition: MenuPosition.below,
-      optionBuilder: (context, option, onOptionSelected) {
-        return GestureDetector(
-          onTap: () {
-            onOptionSelected(option);
-            setState(() {
-              selectedOption = option;
-            });
-          },
-          child: selectedOption == option
-              ? Container(
-                  height: 60,
-                  color: Colors.deepOrange.withOpacity(0.2),
-                  child: Center(
-                    child: Text(
-                      option.label,
-                      style: const TextStyle(
-                        color: Colors.deepOrange,
-                      ),
-                    ),
-                  ),
-                )
-              : SizedBox(
-                  height: 60,
-                  child: Center(
-                    child: Text(
-                      option.label,
-                      style: TextStyle(
-                        color: Colors.orange[700],
-                      ),
-                    ),
-                  ),
-                ),
-        );
-      },
-      menuDecoration: MenuDecoration(
-        backgroundDecoration: BoxDecoration(
-          color: Colors.white,
-          border: Border.all(color: Colors.deepOrange),
-          borderRadius: BorderRadius.circular(12),
-          boxShadow: [
-            BoxShadow(
-              offset: const Offset(0, 1),
-              color: Colors.red[300]!,
-            ),
-          ],
-        ),
-        separatorBuilder: (context, index) => Container(
-          height: 1,
-          width: double.infinity,
-          color: Colors.deepOrange,
-        ),
-      ),
-      inputStyle: const TextStyle(color: Colors.deepOrange),
-      inputDecoration: InputDecoration(
-        hintText: 'Select fruit',
-        filled: true,
-        fillColor: Colors.white,
-        focusColor: Colors.deepOrange,
-        contentPadding: const EdgeInsets.all(16),
-        errorStyle: TextStyle(
-          fontSize: 16,
-          fontWeight: FontWeight.w400,
-          color: Colors.red[900],
-        ),
-        hintStyle: TextStyle(
-          fontSize: 16,
-          fontWeight: FontWeight.w400,
-          color: Colors.orange[700],
-        ),
-        enabledBorder: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: Colors.deepOrange[300]!,
-            width: 1,
-          ),
-          borderRadius: BorderRadius.circular(8),
-        ),
-        border: OutlineInputBorder(
-          borderSide: BorderSide.none,
-          borderRadius: BorderRadius.circular(8),
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderSide: const BorderSide(
-            color: Colors.deepOrange,
-            width: 2,
-          ),
-          borderRadius: BorderRadius.circular(8),
-        ),
-        focusedErrorBorder: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: Colors.red[900]!,
-            width: 2,
-          ),
-          borderRadius: BorderRadius.circular(8),
-        ),
-        errorBorder: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: Colors.red[900]!,
-            width: 2,
-          ),
-          borderRadius: BorderRadius.circular(8),
-        ),
-        disabledBorder: OutlineInputBorder(
-          borderSide: const BorderSide(
-            color: Colors.grey,
-          ),
-          borderRadius: BorderRadius.circular(8),
-        ),
-      ),
-    );
-  }
-}
-
-class MultiSelectDemo<T> extends StatelessWidget {
-  final List<Option<T>> options;
-  const MultiSelectDemo({
-    super.key,
-    required this.options,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return MultiSelectField<T>(
-      options: options,
-      fieldText: 'Select fruit',
     );
   }
 }
