+++ lib/select_field.dart
@@ -5,4 +5,4 @@ export 'src/option.dart';
 export 'src/menu_decoration.dart';
 export 'src/select_field/select_field_menu_controller.dart';
 export 'src/multi_select_field/multi_select_field.dart';
-export 'src/multi_select_field/multi_select_field_menu_controller.dart';
+export 'src/multi_select_field/multi_select_custom_control.dart';
