+++ lib/src/multi_select_field/multi_select_custom_control.dart
@@ -1,0 +1,9 @@
+import 'package:select_field/select_field.dart';
+
+class MultiSelectCustomControl<T> {
+  List<Option<T>> selectedOptions;
+
+  MultiSelectCustomControl({
+    required this.selectedOptions,
+  });
+}
